+++ video_process/video_create.py
@@ -1,0 +1,19 @@
+import subprocess
+
+def create_video(images_path, video_path):
+    command = [
+        'ffmpeg',
+        '-framerate', '30',
+        '-pattern_type', 'glob',
+        '-i', f"{images_path}/*.png",
+        '-c:v', 'libx264',
+        '-pix_fmt', 'yuv420p',
+        video_path
+    ]
+    process = subprocess.Popen(command, stderr=subprocess.PIPE, text=True)
+    returncode = process.wait()  # Wait for the process to complete
+    if returncode != 0:
+        stderr_output = process.stderr.read()
+        error_message = f'ffmpeg error: {stderr_output}'
+        print(error_message)
+        raise Exception(error_message)  # Raise an exception with the error output
